backend/requirements.txt
fastapi
uvicorn
spacy>=3.7.2
sentence-transformers
scikit-learn
networkx
python-multipart
transformers
torch --index-url https://download.pytorch.org/whl/cpu
numpy<2.0.0
pydantic
docx2txt
PyPDF2
python-docx
